Gate embedded services by Go build tag:
The embedded Kube-apiserver, controller manager,
and etcd server add more than double to the size
of the final binary and a log of build dependencies.
This gates all that behind an "embedded" Go build tag.
